Update dependencies
This PR pins to `tchannel-go` 1.x and `zap` 1.x. These libraries are stable
(TChannel hasn't had a breaking change in more than a year), so it doesn't make
sense for this library to lock to minors - it was already three releases behind
the latest TChannel, which creates version conflicts for application owners.