Add a structured way to extract traces for logging (#118)
Since opentracing implementations differ in how they represent tracing data
(e.g., not all implementation use span and trace IDs), there isn't a generic way
to add tracing information to log messages. This PR adds a small utility to
Jaeger's zap integration that extracts structured data from a context.Context.

This is an exact clone of the code currently in UberFX's `ulog` package; for
easy re-usability, I think that this is a better place for it. I'm happy to
remove the code in UberFX once this lands.