Fix FollowsFrom spans ignoring baggage/debug header
This patch fixes the StartSpan() logic so that spans started with
a FollowsFrom reference to a dummy parent (i.e., a parent used for
carrying baggage items or debug ID extracted from HTTP headers)
are set with the correct baggage items or debug ID, respectively.
Previously, these headers were in effect only if the span was
started with a ChildOf reference to the dummy header.

The existing unit tests for baggage items / debug ID headers were
also extended to repeat the test for both ChildOf and FollowsFrom
reference types.

Signed-off-by: Zvi Cahana <zvic@il.ibm.com>