Encode SpanID and ParentSpanID as a 16 hex characters in zipkin propagator
Based on the openzipkin/b3-propagation specification, the SpanID and
ParentSpanID are to be encoded as 16 lower-hex characters.

https://github.com/jaegertracing/jaeger-client-go/pull/533 fixed this
for TraceID, and fixed the SpanID.String() method, but did not update
the Zipkin Propagator SpanID encoding.

This fix uses SpanID.String() to ensure proper encoding of the SpanID and
ParentSpandID within the Zipkin Propagator.

Signed-off-by: Nathan Bowler <nzb301@psu.edu>